+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.4
+
+- Update packages.
+- Bump dart version to 3.0.0.
+- Add number extensions for `num` to format number to localized numbers String like arabic numbers.
+- Add `String` extensions to check if string is rtl or is of specifc language.
+
 ## 0.0.2
 
 - Update packages.